requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chromadb==1.2.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
durationpy==0.10
executing==2.2.1
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kubernetes==34.1.0
langchain==1.0.0
langchain-chroma==1.0.0
langchain-community==0.3.31
langchain-core==1.0.0
langchain-openai==1.0.0
langchain-text-splitters==0.3.11
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.8.0
nest-asyncio==1.6.0
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.1
openai==2.3.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
